Get Protocol Variables for Standardisation

This function identifies which among the potential protocol variables should be used for standardising a given exposure. It exports outputs for the different steps of the process.

Usage

get_protocol_var(data, var_to_std, protocol_vars, covariates, folder, group)

Arguments

data

A data frame containing the exposure and protocol variables.

var_to_std

A character string with the name of the variable to be standardised.

protocol_vars

A character vector of names of potential protocol variables.

covariates

A character vector of names of model covariates.

folder

The directory folder where the output CSV files will be saved.

group

The group/exposure under consideration.

Value

A character vector containing the names of the final protocol variables to be used for standardisation.